+++ hospital-appointments/src/main/java/ba/unsa/etf/hospital/service/RoleService.java
@@ -1,5 +1,6 @@
 package ba.unsa.etf.hospital.service;
 
+import ba.unsa.etf.hospital.exception.RoleNotFoundException;
 import ba.unsa.etf.hospital.model.Role;
 import ba.unsa.etf.hospital.repository.RoleRepository;
 import org.springframework.stereotype.Service;
@@ -10,12 +11,15 @@ import java.util.Optional;
 @Service
 public class RoleService {
     private final RoleRepository roleRepository;
+
     public RoleService(RoleRepository roleRepository) {
         this.roleRepository = roleRepository;
     }
+
     public List<Role> getAllRoles() {
         return roleRepository.findAll();
     }
+
     public Role saveRole(Role role) {
         return roleRepository.save(role);
     }
@@ -25,6 +29,9 @@ public class RoleService {
     }
 
     public void deleteById(Long id) {
+        if (!roleRepository.existsById(id)) {
+            throw new RoleNotFoundException(id);
+        }
         roleRepository.deleteById(id);
     }
 }
