+++ hospital-appointments/src/main/java/ba/unsa/etf/hospital/service/TerminService.java
@@ -1,5 +1,6 @@
 package ba.unsa.etf.hospital.service;
 
+import ba.unsa.etf.hospital.exception.TerminNotFoundException;
 import ba.unsa.etf.hospital.model.Termin;
 import ba.unsa.etf.hospital.repository.TerminRepository;
 import org.springframework.stereotype.Service;
@@ -10,12 +11,15 @@ import java.util.Optional;
 @Service
 public class TerminService {
     private final TerminRepository terminRepository;
+
     public TerminService(TerminRepository terminRepository) {
         this.terminRepository = terminRepository;
     }
+
     public List<Termin> getAllTermini(){
         return terminRepository.findAll();
     }
+
     public Termin saveTermin(Termin termin) {
         return terminRepository.save(termin);
     }
@@ -25,6 +29,9 @@ public class TerminService {
     }
 
     public void deleteById(Long id) {
+        if (!terminRepository.existsById(id)) {
+            throw new TerminNotFoundException(id);
+        }
         terminRepository.deleteById(id);
     }
 }
